+++ libretro/libretro-mapper.c
@@ -16,7 +16,7 @@
 #include "disk.h"
 #include "hrtimer.h"
 
-static retro_input_state_t input_state_cb;
+retro_input_state_t input_state_cb;
 static retro_input_poll_t input_poll_cb;
 
 void retro_set_input_state(retro_input_state_t cb)
@@ -29,20 +29,6 @@ void retro_set_input_poll(retro_input_poll_t cb)
    input_poll_cb = cb;
 }
 
-#ifdef POINTER_DEBUG
-int pointer_x = 0;
-int pointer_y = 0;
-#endif
-int last_pointer_x = 0;
-int last_pointer_y = 0;
-
-int vkey_pos_x = 0;
-int vkey_pos_y = 0;
-int vkbd_x_min = 0;
-int vkbd_x_max = 0;
-int vkbd_y_min = 0;
-int vkbd_y_max = 0;
-
 /* Mouse speed flags */
 #define MOUSE_SPEED_SLOWER 1
 #define MOUSE_SPEED_FASTER 2
@@ -54,14 +40,14 @@ int vkbd_y_max = 0;
 
 /* Core flags */
 int mapper_keys[RETRO_MAPPER_LAST] = {0};
-int vkflag[10] = {0};
 int retro_capslock = false;
 bool retro_mousemode = false;
 bool mousemode_locked = false;
-static int jflag[4][16] = {0};
-static int kjflag[2][16] = {0};
-static int mflag[2][16] = {0};
-static int aflag[2][16] = {0};
+
+int jflag[4][16] = {0};
+int kjflag[2][16] = {0};
+int mflag[2][16] = {0};
+int aflag[2][16] = {0};
 static int mapper_flag[4][16] = {0};
 static unsigned int mouse_speed[2] = {0};
 
@@ -69,26 +55,18 @@ extern bool request_update_av_info;
 extern void retro_reset_soft();
 extern bool retro_statusbar;
 extern bool retro_vkbd;
-extern bool retro_vkbd_page;
-extern bool retro_vkbd_transparent;
 extern short int retro_vkbd_ready;
-extern void retro_key_up(int);
-extern void retro_key_down(int);
-extern void retro_mouse(int, int, int);
-extern void retro_mouse_button(int, int, int);
-extern void retro_joystick(int, int, int);
-extern void retro_joystick_analog(int, int, int);
-extern void retro_joystick_button(int, int, int);
-extern unsigned int retro_devices[RETRO_DEVICES];
+extern void input_vkbd(void);
+extern int vkflag[10];
 
 static unsigned retro_key_state[RETROK_LAST] = {0};
 static unsigned retro_key_event_state[RETROK_LAST] = {0};
-static int16_t joypad_bits[RETRO_DEVICES];
+int16_t joypad_bits[RETRO_DEVICES];
 extern bool libretro_supports_bitmasks;
 extern dc_storage *dc;
+extern unsigned char statusbar_text[RETRO_PATH_MAX];
 
 /* Core options */
-extern unsigned int video_config;
 extern unsigned int video_config_aspect;
 extern bool opt_aspect_ratio_locked;
 extern unsigned int zoom_mode_id;
@@ -100,14 +78,8 @@ extern unsigned int opt_analogmouse_deadzone;
 extern float opt_analogmouse_speed;
 extern unsigned int opt_dpadmouse_speed;
 extern unsigned int opt_physicalmouse;
-extern bool opt_keyrah_keypad;
 extern bool opt_keyboard_pass_through;
-extern int opt_cd32pad_options;
-extern int opt_retropad_options;
 extern char opt_joyport_order[5];
-
-int imagename_timer = 0;
-static unsigned char statusbar_text[RETRO_PATH_MAX] = {0};
 
 bool retro_turbo_fire = false;
 bool turbo_fire_locked = false;
@@ -115,36 +87,6 @@ unsigned int turbo_fire_button = 0;
 unsigned int turbo_pulse = 6;
 unsigned int turbo_state[RETRO_DEVICES] = {0};
 unsigned int turbo_toggle[RETRO_DEVICES] = {0};
-
-enum EMU_FUNCTIONS
-{
-   EMU_VKBD = 0,
-   EMU_STATUSBAR,
-   EMU_JOYMOUSE,
-   EMU_RESET,
-   EMU_SAVE_DISK,
-   EMU_ASPECT_RATIO,
-   EMU_ZOOM_MODE,
-   EMU_TURBO_FIRE,
-   EMU_FUNCTION_COUNT
-};
-
-/* VKBD_MIN_HOLDING_TIME: Hold a direction longer than this and automatic movement sets in */
-/* VKBD_MOVE_DELAY: Delay between automatic movement from button to button */
-#define VKBD_MIN_HOLDING_TIME 200
-#define VKBD_MOVE_DELAY 50
-bool let_go_of_direction = true;
-long last_move_time = 0;
-long last_press_time = 0;
-
-/* VKBD_STICKY_HOLDING_TIME: Button press longer than this triggers sticky key */
-#define VKBD_STICKY_HOLDING_TIME 1000
-int let_go_of_button = 1;
-long last_press_time_button = 0;
-int vkey_pressed = -1;
-int vkey_sticky = -1;
-int vkey_sticky1 = -1;
-int vkey_sticky2 = -1;
 
 void emu_function(int function)
 {
@@ -211,484 +153,6 @@ void emu_function(int function)
          dc_save_disk_toggle(dc, false, true);
          break;
    }
-}
-
-static unsigned char* joystick_value_human(int val[16], int uae_device)
-{
-   /*
-    * uae_device:
-    * 0 = Single button joystick (Keyrah + Parallel)
-    * 1 = RetroPad
-    * 2 = CD32 Pad
-    * 3 = Mouse
-    * 4 = Analog joystick
-    */
-
-   unsigned str_len = 4;
-   unsigned char *str = malloc(sizeof(char)*str_len);
-   snprintf(str, sizeof(unsigned char)*str_len, "%3s", "   ");
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_UP] || val[RETRO_DEVICE_ID_JOYPAD_SELECT]) /* Unused SELECT acts as a jump button */
-      str[1] = 30;
-
-   else if (val[RETRO_DEVICE_ID_JOYPAD_DOWN])
-      str[1] = 28;
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_LEFT])
-      str[0] = 27;
-
-   else if (val[RETRO_DEVICE_ID_JOYPAD_RIGHT])
-      str[2] = 29;
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_B])
-   {
-      switch (uae_device)
-      {
-         case 1:
-            if (opt_retropad_options == 1 || opt_retropad_options == 3)
-               str[1] = ('2' | 0x80);
-            else
-               str[1] = ('1' | 0x80);
-            break;
-         case 3:
-            str[1] = ('L' | 0x80);
-            break;
-         case 4:
-            str[1] = ('1' | 0x80);
-            break;
-         default:
-            str[1] = (str[1] | 0x80);
-            break;
-      }
-   }
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_A])
-   {
-      switch (uae_device)
-      {
-         case 1:
-            if (opt_retropad_options == 1 || opt_retropad_options == 3)
-               ; /* no-op */
-            else
-               str[1] = ('2' | 0x80);
-            break;
-         case 3:
-            str[1] = ('R' | 0x80);
-            break;
-         case 4:
-            str[1] = ('2' | 0x80);
-            break;
-         default:
-            str[1] = (str[1] | 0x80);
-            break;
-      }
-   }
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_Y])
-   {
-      switch (uae_device)
-      {
-         case 1:
-            if (opt_retropad_options == 1 || opt_retropad_options == 3)
-               str[1] = ('1' | 0x80);
-            break;
-         case 3:
-            str[1] = ('M' | 0x80);
-            break;
-         case 4:
-            str[1] = ('3' | 0x80);
-            break;
-         default:
-            str[1] = (str[1] | 0x80);
-            break;
-      }
-   }
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_X])
-   {
-      switch (uae_device)
-      {
-         case 4:
-            str[1] = ('4' | 0x80);
-            break;
-         default:
-            str[1] = (str[1] | 0x80);
-            break;
-      }
-   }
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_START])
-   {
-      switch (uae_device)
-      {
-         case 2:
-            str[1] = ('P' | 0x80);
-            break;
-      }
-   }
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_L])
-   {
-      switch (uae_device)
-      {
-         case 2:
-            str[0] = ('R' | 0x80);
-            break;
-      }
-   }
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_R])
-   {
-      switch (uae_device)
-      {
-         case 2:
-            str[2] = ('F' | 0x80);
-            break;
-      }
-   }
-
-   return str;
-}
-
-static unsigned int joystick_color(int val[16])
-{
-   unsigned color = 0;
-
-   if (opt_cd32pad_options == 1 || opt_cd32pad_options == 3)
-   {
-      if (val[RETRO_DEVICE_ID_JOYPAD_Y])
-         color |= (pix_bytes == 4) ? RGB888(248,0,0) : RGB565(255,0,0);
-
-      if (val[RETRO_DEVICE_ID_JOYPAD_B])
-         color |= (pix_bytes == 4) ? RGB888(0,0,248) : RGB565(0,0,255);
-
-      if (val[RETRO_DEVICE_ID_JOYPAD_X])
-         color |= (pix_bytes == 4) ? RGB888(0,248,0) : RGB565(0,255,0);
-
-      if (val[RETRO_DEVICE_ID_JOYPAD_A])
-         color |= (pix_bytes == 4) ? RGB888(248,248,0) : RGB565(255,255,0);
-   }
-   else
-   {
-      if (val[RETRO_DEVICE_ID_JOYPAD_B])
-         color |= (pix_bytes == 4) ? RGB888(248,0,0) : RGB565(255,0,0);
-
-      if (val[RETRO_DEVICE_ID_JOYPAD_A])
-         color |= (pix_bytes == 4) ? RGB888(0,0,248) : RGB565(0,0,255);
-
-      if (val[RETRO_DEVICE_ID_JOYPAD_Y])
-         color |= (pix_bytes == 4) ? RGB888(0,248,0) : RGB565(0,255,0);
-
-      if (val[RETRO_DEVICE_ID_JOYPAD_X])
-         color |= (pix_bytes == 4) ? RGB888(248,248,0) : RGB565(255,255,0);
-   }
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_L])
-      color |= (pix_bytes == 4) ? RGB888(170,170,170) : RGB565(110,110,110);
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_R])
-      color |= (pix_bytes == 4) ? RGB888(170,170,170) : RGB565(110,110,110);
-
-   if (val[RETRO_DEVICE_ID_JOYPAD_START])
-      color |= (pix_bytes == 4) ? RGB888(164,164,164) : RGB565(72,72,72);
-
-   if (color == 0)
-      color = (pix_bytes == 4) ? 0xffffff : 0xffff;
-
-   return color;
-}
-
-static bool flag_empty(int val[16])
-{
-   for (int x = 0; x < 16; x++)
-   {
-      if (val[x])
-         return false;
-   }
-   return true;
-}
-
-void display_current_image(const char *image, bool inserted)
-{
-   static char imagename[RETRO_PATH_MAX] = {0};
-   static char imagename_prev[RETRO_PATH_MAX] = {0};
-
-   imagename_timer = 150;
-   if (strcmp(image, ""))
-   {
-      snprintf(imagename, sizeof(imagename), "%s%.98s", "  ", path_basename(image));
-      snprintf(imagename_prev, sizeof(imagename_prev), "%.100s", imagename);
-   }
-   else
-      snprintf(imagename, sizeof(imagename), "%.100s", imagename_prev);
-
-   snprintf(&statusbar_text[0], sizeof(statusbar_text), "%-100s", imagename);
-
-   if (inserted)
-      statusbar_text[0] = (8 | 0x80);
-   else if (!strcmp(image, ""))
-      statusbar_text[0] = (9 | 0x80);
-}
-
-void print_statusbar(void)
-{
-   if (opt_statusbar & STATUSBAR_BASIC && !imagename_timer)
-      return;
-
-   int BOX_Y                = 0;
-   int BOX_WIDTH            = 0;
-   int BOX_HEIGHT           = 11;
-   int BOX_PADDING          = 2;
-
-   int FONT_WIDTH           = 1;
-   if (video_config & PUAE_VIDEO_HIRES)
-   {
-      if (video_config & PUAE_VIDEO_DOUBLELINE)
-         FONT_WIDTH         = 1;
-      else
-         FONT_WIDTH         = 2;
-   }
-   else if (video_config & PUAE_VIDEO_SUPERHIRES)
-   {
-      if (video_config & PUAE_VIDEO_DOUBLELINE)
-         FONT_WIDTH         = 2;
-      else
-         FONT_WIDTH         = 4;
-   }
-   int FONT_HEIGHT          = 1;
-   int FONT_COLOR           = (pix_bytes == 4) ? 0xffffff : 0xffff;;
-   int FONT_SLOT            = 34 * FONT_WIDTH;
-
-   int TEXT_X               = 1 * FONT_WIDTH;
-   int TEXT_Y               = 0;
-   int TEXT_LENGTH          = (video_config & PUAE_VIDEO_DOUBLELINE) ? 128 : 64;
-
-   /* Statusbar location */
-   /* Top */
-   if (opt_statusbar_position < 0)
-      TEXT_Y = BOX_PADDING;
-   /* Bottom */
-   else
-      TEXT_Y = gfxvidinfo.outheight - opt_statusbar_position - BOX_HEIGHT + BOX_PADDING;
-   BOX_Y = TEXT_Y - BOX_PADDING;
-
-   /* Statusbar size */
-   BOX_WIDTH = zoomed_width;
-   int ZOOMED_WIDTH_OFFSET = retrow - zoomed_width;
-
-   /* Video resolution */
-   int TEXT_X_RESOLUTION = TEXT_X + (FONT_SLOT*4) + (FONT_WIDTH*16) - (ZOOMED_WIDTH_OFFSET/2);
-   unsigned char RESOLUTION[10] = {0};
-   snprintf(RESOLUTION, sizeof(RESOLUTION), "%4dx%3d", zoomed_width, zoomed_height);
-
-   /* Model & memory */
-   int TEXT_X_MODEL  = TEXT_X + (FONT_SLOT*6) + (FONT_WIDTH*32) - ZOOMED_WIDTH_OFFSET;
-   int TEXT_X_MEMORY = TEXT_X + (FONT_SLOT*6) + (FONT_WIDTH*6) - ZOOMED_WIDTH_OFFSET;
-   /* Sacrifice memory slot if there is not enough width */
-   if (!(video_config & PUAE_VIDEO_DOUBLELINE))
-   {
-      if (TEXT_X_MEMORY < (TEXT_X_RESOLUTION + FONT_SLOT + (FONT_WIDTH*20)))
-         TEXT_X_MEMORY = -1;
-   }
-
-   unsigned char MODEL[10] = {0};
-   unsigned char MEMORY[5] = {0};
-   float mem_size = 0;
-   mem_size  = (float)(currprefs.chipmem_size / 0x80000) / 2;
-   mem_size += (float)(currprefs.bogomem_size / 0x40000) / 4;
-   mem_size += (float)(currprefs.fastmem_size / 0x100000);
-   if (TEXT_X_MEMORY > 0)
-      snprintf(MEMORY, sizeof(MEMORY), "%2.0fM", floor(mem_size));
-   switch (currprefs.cs_compatible)
-   {
-      case CP_A500:
-         snprintf(MODEL, sizeof(MODEL), "%s", " A500");
-         break;
-      case CP_A500P:
-         snprintf(MODEL, sizeof(MODEL), "%s", "A500+");
-         break;
-      case CP_A600:
-         snprintf(MODEL, sizeof(MODEL), "%s", " A600");
-         break;
-      case CP_A1200:
-         snprintf(MODEL, sizeof(MODEL), "%s", "A1200");
-         break;
-      case CP_A4000:
-         snprintf(MODEL, sizeof(MODEL), "%s", "A4000");
-         break;
-      case CP_CDTV:
-         snprintf(MODEL, sizeof(MODEL), "%s", " CDTV");
-         break;
-      case CP_CD32:
-         snprintf(MODEL, sizeof(MODEL), "%s", " CD32");
-         break;
-   }
-
-   /* Double line positions */
-   if (video_config & PUAE_VIDEO_DOUBLELINE)
-   {
-      TEXT_X_RESOLUTION = TEXT_X + (FONT_SLOT*15) + (FONT_WIDTH*5) - ZOOMED_WIDTH_OFFSET;
-      TEXT_X_MODEL      = TEXT_X + (FONT_SLOT*17) + (FONT_WIDTH*15) - ZOOMED_WIDTH_OFFSET;
-      TEXT_X_MEMORY     = TEXT_X + (FONT_SLOT*16) + (FONT_WIDTH*25) - ZOOMED_WIDTH_OFFSET;
-   }
-
-   /* Joy port indicators */
-   unsigned char JOYMODE1[5] = {0};
-   unsigned char JOYMODE2[5] = {0};
-   unsigned char JOYMODE3[5] = {0};
-   unsigned char JOYMODE4[5] = {0};
-
-   unsigned char JOYPORT1[5] = {0};
-   unsigned char JOYPORT2[5] = {0};
-   unsigned char JOYPORT3[5] = {0};
-   unsigned char JOYPORT4[5] = {0};
-
-   /* Joy port positions */
-   int TEXT_X_JOYMODE1 = TEXT_X;
-   int TEXT_X_JOYPORT1 = TEXT_X_JOYMODE1 + (FONT_WIDTH*13);
-
-   int TEXT_X_JOYMODE2 = TEXT_X + FONT_SLOT;
-   int TEXT_X_JOYPORT2 = TEXT_X_JOYMODE2 + (FONT_WIDTH*13);
-
-   int TEXT_X_JOYMODE3 = TEXT_X + (FONT_SLOT*2);
-   int TEXT_X_JOYPORT3 = TEXT_X_JOYMODE3 + (FONT_WIDTH*13);
-
-   int TEXT_X_JOYMODE4 = TEXT_X + (FONT_SLOT*3);
-   int TEXT_X_JOYPORT4 = TEXT_X_JOYMODE4 + (FONT_WIDTH*13);
-
-   /* Regular joyflags */
-   if (!retro_mousemode)
-   {
-      switch (retro_devices[0])
-      {
-         case RETRO_DEVICE_PUAE_CD32PAD:
-            snprintf(JOYMODE1, sizeof(JOYMODE1), "%2s", "C1");
-            break;
-         case RETRO_DEVICE_PUAE_ANALOG:
-            snprintf(JOYMODE1, sizeof(JOYMODE1), "%2s", "A1");
-            break;
-         default:
-            snprintf(JOYMODE1, sizeof(JOYMODE1), "%2s", "J1");
-            break;
-      }
-
-      switch (retro_devices[1])
-      {
-         case RETRO_DEVICE_PUAE_CD32PAD:
-            snprintf(JOYMODE2, sizeof(JOYMODE2), "%2s", "C2");
-            break;
-         case RETRO_DEVICE_PUAE_ANALOG:
-            snprintf(JOYMODE1, sizeof(JOYMODE1), "%2s", "A2");
-            break;
-         default:
-            snprintf(JOYMODE2, sizeof(JOYMODE2), "%2s", "J2");
-            break;
-      }
-   }
-   else
-   {
-      snprintf(JOYMODE1, sizeof(JOYMODE1), "%2s", "M1");
-      snprintf(JOYMODE2, sizeof(JOYMODE2), "%2s", "M2");
-   }
-
-   /* Regular ports */
-   switch (retro_devices[0])
-   {
-      case RETRO_DEVICE_PUAE_CD32PAD:
-         snprintf(JOYPORT1, sizeof(JOYPORT1), "%3s", joystick_value_human(jflag[0], 2));
-         break;
-      case RETRO_DEVICE_PUAE_ANALOG:
-         snprintf(JOYPORT1, sizeof(JOYPORT1), "%3s", joystick_value_human(aflag[0], 4));
-         break;
-      default:
-         snprintf(JOYPORT1, sizeof(JOYPORT1), "%3s", joystick_value_human(jflag[0], 1));
-         break;
-   }
-
-   switch (retro_devices[1])
-   {
-      case RETRO_DEVICE_PUAE_CD32PAD:
-         snprintf(JOYPORT2, sizeof(JOYPORT2), "%3s", joystick_value_human(jflag[1], 2));
-         break;
-      case RETRO_DEVICE_PUAE_ANALOG:
-         snprintf(JOYPORT2, sizeof(JOYPORT2), "%3s", joystick_value_human(aflag[1], 4));
-         break;
-      default:
-         snprintf(JOYPORT2, sizeof(JOYPORT2), "%3s", joystick_value_human(jflag[1], 1));
-         break;
-   }
-
-   /* Parallel ports, hidden if not connected */
-   if (retro_devices[2])
-   {
-      snprintf(JOYMODE3, sizeof(JOYMODE3), "%2s", "J3");
-      snprintf(JOYPORT3, sizeof(JOYPORT3), "%3s", joystick_value_human(jflag[2], 0));
-   }
-   if (retro_devices[3])
-   {
-      snprintf(JOYMODE4, sizeof(JOYMODE4), "%2s", "J4");
-      snprintf(JOYPORT4, sizeof(JOYPORT4), "%3s", joystick_value_human(jflag[3], 0));
-   }
-
-   /* Mouse flags */
-   if (!flag_empty(mflag[1]))
-   {
-      snprintf(JOYMODE1, sizeof(JOYMODE1), "%2s", "M1");
-      snprintf(JOYPORT1, sizeof(JOYPORT1), "%3s", joystick_value_human(mflag[1], 3));
-   }
-   if (!flag_empty(mflag[0]))
-   {
-      snprintf(JOYMODE2, sizeof(JOYMODE2), "%2s", "M2");
-      snprintf(JOYPORT2, sizeof(JOYPORT2), "%3s", joystick_value_human(mflag[0], 3));
-   }
-
-   /* Keyrah flags */
-   if (opt_keyrah_keypad)
-   {
-      if (!flag_empty(kjflag[0]))
-      {
-         snprintf(JOYMODE1, sizeof(JOYMODE1), "%2s", "K1");
-         snprintf(JOYPORT1, sizeof(JOYPORT1), "%3s", joystick_value_human(kjflag[0], 0));
-      }
-      if (!flag_empty(kjflag[1]))
-      {
-         snprintf(JOYMODE2, sizeof(JOYMODE2), "%2s", "K2");
-         snprintf(JOYPORT2, sizeof(JOYPORT2), "%3s", joystick_value_human(kjflag[1], 0));
-      }
-   }
-
-   /* Button colorize CD32 Pad */
-   int JOY1_COLOR = FONT_COLOR;
-   int JOY2_COLOR = FONT_COLOR;
-   if (retro_devices[0] == RETRO_DEVICE_PUAE_CD32PAD)
-      JOY1_COLOR = joystick_color(jflag[0]);
-   if (retro_devices[1] == RETRO_DEVICE_PUAE_CD32PAD)
-      JOY2_COLOR = joystick_color(jflag[1]);
-
-   /* Statusbar output */
-   draw_fbox(0, BOX_Y, BOX_WIDTH, BOX_HEIGHT, 0, GRAPH_ALPHA_100);
-
-   if (imagename_timer > 0)
-   {
-      draw_text(TEXT_X, TEXT_Y, FONT_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, TEXT_LENGTH, statusbar_text);
-      return;
-   }
-   
-   draw_text(TEXT_X_JOYMODE1,   TEXT_Y, FONT_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, 10, JOYMODE1);
-   draw_text(TEXT_X_JOYPORT1,   TEXT_Y, JOY1_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, 10, JOYPORT1);
-
-   draw_text(TEXT_X_JOYMODE2,   TEXT_Y, FONT_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, 10, JOYMODE2);
-   draw_text(TEXT_X_JOYPORT2,   TEXT_Y, JOY2_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, 10, JOYPORT2);
-
-   draw_text(TEXT_X_JOYMODE3,   TEXT_Y, FONT_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, 10, JOYMODE3);
-   draw_text(TEXT_X_JOYPORT3,   TEXT_Y, FONT_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, 10, JOYPORT3);
-
-   draw_text(TEXT_X_JOYMODE4,   TEXT_Y, FONT_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, 10, JOYMODE4);
-   draw_text(TEXT_X_JOYPORT4,   TEXT_Y, FONT_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, 10, JOYPORT4);
-
-   draw_text(TEXT_X_RESOLUTION, TEXT_Y, FONT_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, 10, RESOLUTION);
-   draw_text(TEXT_X_MEMORY,     TEXT_Y, FONT_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, 10, MEMORY);
-   draw_text(TEXT_X_MODEL,      TEXT_Y, FONT_COLOR, 0, GRAPH_ALPHA_100, GRAPH_BG_ALL, FONT_WIDTH, FONT_HEIGHT, 10, MODEL);
 }
 
 static void process_keyrah()
@@ -1305,83 +769,18 @@ void update_input(unsigned disable_keys)
     * INDEX  0  1  2  3  4  5  6  7  8  9  10  11  12  13  14  15  16  17  18  19  20  21  22  23
     */
 
-   static long now = 0;
-   static long last_vkey_pressed_time = 0;
-   static int last_vkey_pressed = -1;
-   static int vkey_sticky1_release = 0;
-   static int vkey_sticky2_release = 0;
-
-   static int i = 0, j = 0, mk = 0;
-   static int LX = 0, LY = 0, RX = 0, RY = 0;
+   unsigned int i = 0, j = 0, mk = 0;
+   int LX = 0, LY = 0, RX = 0, RY = 0;
    const int threshold = 20000;
 
-   static int jbt[2][24] = {0};
+   static int jbt[2][RETRO_DEVICE_ID_JOYPAD_LAST] = {0};
    static int kbt[EMU_FUNCTION_COUNT] = {0};
 
-   now = retro_ticks() / 1000;
-
-   if (vkey_sticky && last_vkey_pressed != -1 && last_vkey_pressed > 0)
-   {
-      if (vkey_sticky1 > -1 && vkey_sticky1 != last_vkey_pressed)
-      {
-         if (vkey_sticky2 > -1 && vkey_sticky2 != last_vkey_pressed)
-            retro_key_up(vkey_sticky2);
-         vkey_sticky2 = last_vkey_pressed;
-      }
-      else
-         vkey_sticky1 = last_vkey_pressed;
-   }
-
-   /* Keyup only after button is up */
-   if (last_vkey_pressed != -1 && !vkflag[RETRO_DEVICE_ID_JOYPAD_B])
-   {
-      if (last_vkey_pressed < -10)
-      {
-         if (last_vkey_pressed == -15)
-         {
-            retro_mouse_button(0, 0, 0);
-            mflag[0][RETRO_DEVICE_ID_JOYPAD_B] = 0;
-         }
-         else if (last_vkey_pressed == -16)
-         {
-            retro_mouse_button(0, 1, 0);
-            mflag[0][RETRO_DEVICE_ID_JOYPAD_A] = 0;
-         }
-         else if (last_vkey_pressed == -17)
-         {
-            retro_mouse_button(0, 2, 0);
-            mflag[0][RETRO_DEVICE_ID_JOYPAD_Y] = 0;
-         }
-      }
-      else
-      {
-         if (vkey_pressed == -1 && last_vkey_pressed >= 0 && last_vkey_pressed != vkey_sticky1 && last_vkey_pressed != vkey_sticky2)
-            retro_key_up(last_vkey_pressed);
-
-         if (retro_capslock)
-            retro_key_up(keyboard_translation[RETROK_LSHIFT]);
-      }
-
-      last_vkey_pressed = -1;
-   }
-
-   if (vkey_sticky1_release)
-   {
-      vkey_sticky1_release = 0;
-      vkey_sticky1 = -1;
-      retro_key_up(vkey_sticky1);
-   }
-   if (vkey_sticky2_release)
-   {
-      vkey_sticky2_release = 0;
-      vkey_sticky2 = -1;
-      retro_key_up(vkey_sticky2);
-   }
-
    /* Keyboard hotkeys */
-   for (i = 0; i < EMU_FUNCTION_COUNT; i++)
+   for (i = 0; i < (RETRO_MAPPER_LAST - RETRO_DEVICE_ID_JOYPAD_LAST); i++)
    {
-      mk = i + 24;
+      /* Skip RetroPad mappings from mapper_keys */
+      mk = i + RETRO_DEVICE_ID_JOYPAD_LAST;
 
       /* Key down */
       if (input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, mapper_keys[mk]) && !kbt[i] && mapper_keys[mk])
@@ -1668,381 +1067,7 @@ void update_input(unsigned disable_keys)
    } /* for j */
 
    /* Virtual keyboard for ports 1 & 2 */
-   if (retro_vkbd)
-   {
-      /* Wait for all inputs to be released */
-      if (retro_vkbd_ready < 1)
-      {
-         if ((retro_vkbd_ready == 0 && !joypad_bits[0] && !joypad_bits[1]) ||
-              retro_vkbd_ready < 0)
-            retro_vkbd_ready++;
-         return;
-      }
-
-      /* VKBD Mouse acceleration */
-      const int mspeed_default = 3;
-      static int mspeed;
-      if (!vkflag[RETRO_DEVICE_ID_JOYPAD_B])
-         mspeed = mspeed_default;
-
-      if (!vkflag[RETRO_DEVICE_ID_JOYPAD_B]) /* Allow directions when key is not pressed */
-      {
-         if (!vkflag[RETRO_DEVICE_ID_JOYPAD_UP] && ((joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_UP)) ||
-                                                    (joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_UP)) ||
-                                                    input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_UP)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_UP] = 1;
-         else
-         if (vkflag[RETRO_DEVICE_ID_JOYPAD_UP] && (!(joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_UP)) &&
-                                                   !(joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_UP)) &&
-                                                   !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_UP)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_UP] = 0;
-
-         if (!vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN] && ((joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_DOWN)) ||
-                                                      (joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_DOWN)) ||
-                                                      input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_DOWN)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN] = 1;
-         else
-         if (vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN] && (!(joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_DOWN)) &&
-                                                     !(joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_DOWN)) &&
-                                                     !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_DOWN)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN] = 0;
-
-         if (!vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT] && ((joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_LEFT)) ||
-                                                      (joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_LEFT)) ||
-                                                      input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_LEFT)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT] = 1;
-         else
-         if (vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT] && (!(joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_LEFT)) &&
-                                                     !(joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_LEFT)) &&
-                                                     !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_LEFT)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT] = 0;
-
-         if (!vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT] && ((joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_RIGHT)) ||
-                                                       (joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_RIGHT)) ||
-                                                       input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_RIGHT)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT] = 1;
-         else
-         if (vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT] && (!(joypad_bits[0] & (1 << RETRO_DEVICE_ID_JOYPAD_RIGHT)) &&
-                                                      !(joypad_bits[1] & (1 << RETRO_DEVICE_ID_JOYPAD_RIGHT)) &&
-                                                      !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_RIGHT)))
-            vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT] = 0;
-      }
-      else /* Release all directions when key is pressed */
-      {
-         vkflag[RETRO_DEVICE_ID_JOYPAD_UP]    = 0;
-         vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN]  = 0;
-         vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT]  = 0;
-         vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT] = 0;
-      }
-
-      if (vkflag[RETRO_DEVICE_ID_JOYPAD_UP] ||
-          vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN] ||
-          vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT] ||
-          vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT])
-      {
-         if (let_go_of_direction)
-            /* just pressing down */
-            last_press_time = now;
-
-         if ((now - last_press_time > VKBD_MIN_HOLDING_TIME
-           && now - last_move_time > VKBD_MOVE_DELAY)
-           || let_go_of_direction)
-         {
-            last_move_time = now;
-
-            if (vkflag[RETRO_DEVICE_ID_JOYPAD_UP])
-               vkey_pos_y -= 1;
-            else if (vkflag[RETRO_DEVICE_ID_JOYPAD_DOWN])
-               vkey_pos_y += 1;
-
-            if (vkflag[RETRO_DEVICE_ID_JOYPAD_LEFT])
-               vkey_pos_x -= 1;
-            else if (vkflag[RETRO_DEVICE_ID_JOYPAD_RIGHT])
-               vkey_pos_x += 1;
-         }
-         let_go_of_direction = false;
-      }
-      else
-         let_go_of_direction = true;
-
-      if (vkey_pos_x < 0)
-         vkey_pos_x = VKBDX - 1;
-      else if (vkey_pos_x > VKBDX - 1)
-         vkey_pos_x = 0;
-      if (vkey_pos_y < 0)
-         vkey_pos_y = VKBDY - 1;
-      else if (vkey_pos_y > VKBDY - 1)
-         vkey_pos_y = 0;
-
-      /* Absolute pointer */
-      int p_x = input_state_cb(0, RETRO_DEVICE_POINTER, 0, RETRO_DEVICE_ID_POINTER_X);
-      int p_y = input_state_cb(0, RETRO_DEVICE_POINTER, 0, RETRO_DEVICE_ID_POINTER_Y);
-
-      if (p_x != 0 && p_y != 0 && (p_x != last_pointer_x || p_y != last_pointer_y))
-      {
-         int px = (int)((p_x + 0x7fff) * zoomed_width / 0xffff);
-         int py = (int)((p_y + 0x7fff) * zoomed_height / 0xffff);
-         last_pointer_x = p_x;
-         last_pointer_y = p_y;
-#ifdef POINTER_DEBUG
-         pointer_x = px;
-         pointer_y = py;
-#endif
-         if (px >= vkbd_x_min && px <= vkbd_x_max && py >= vkbd_y_min && py <= vkbd_y_max)
-         {
-            float vkey_width = (float)(vkbd_x_max - vkbd_x_min) / VKBDX;
-            vkey_pos_x = ((px - vkbd_x_min) / vkey_width);
-
-            float vkey_height = (float)(vkbd_y_max - vkbd_y_min) / VKBDY;
-            vkey_pos_y = ((py - vkbd_y_min) / vkey_height);
-
-            vkey_pos_x = (vkey_pos_x < 0) ? 0 : vkey_pos_x;
-            vkey_pos_x = (vkey_pos_x > VKBDX - 1) ? VKBDX - 1 : vkey_pos_x;
-            vkey_pos_y = (vkey_pos_y < 0) ? 0 : vkey_pos_y;
-            vkey_pos_y = (vkey_pos_y > VKBDY - 1) ? VKBDY - 1 : vkey_pos_y;
-
-#ifdef POINTER_DEBUG
-            printf("px:%d py:%d (%d,%d) vkey:%dx%d\n", p_x, p_y, px, py, vkey_pos_x, vkey_pos_y);
-#endif
-         }
-      }
-
-      /* Press Return, RetroPad Start */
-      i = RETRO_DEVICE_ID_JOYPAD_START;
-      if (!vkflag[i] && mapper_keys[i] >= 0 && ((joypad_bits[0] & (1 << i)) ||
-                                                (joypad_bits[1] & (1 << i)))
-                                            && !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_RETURN))
-      {
-         vkflag[i] = 1;
-         retro_key_down(AK_RET);
-      }
-      else
-      if (vkflag[i] && (!(joypad_bits[0] & (1 << i)) &&
-                        !(joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 0;
-         retro_key_up(AK_RET);
-      }
-
-      /* Toggle CapsLock, RetroPad Y */
-      i = RETRO_DEVICE_ID_JOYPAD_Y;
-      if (!vkflag[i] && mapper_keys[i] >= 0 && ((joypad_bits[0] & (1 << i)) ||
-                                                (joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 1;
-         retro_capslock = !retro_capslock;
-      }
-      else
-      if (vkflag[i] && (!(joypad_bits[0] & (1 << i)) &&
-                        !(joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 0;
-      }
-
-      /* Press Space, RetroPad X */
-      i = RETRO_DEVICE_ID_JOYPAD_X;
-      if (!vkflag[i] && mapper_keys[i] >= 0 && ((joypad_bits[0] & (1 << i)) ||
-                                                (joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 1;
-         retro_key_down(AK_SPC);
-      }
-      else
-      if (vkflag[i] && (!(joypad_bits[0] & (1 << i)) &&
-                        !(joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 0;
-         retro_key_up(AK_SPC);
-      }
-
-      /* Toggle transparency, RetroPad A */
-      i = RETRO_DEVICE_ID_JOYPAD_A;
-      if (!vkflag[i] && mapper_keys[i] >= 0 && ((joypad_bits[0] & (1 << i)) ||
-                                                (joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 1;
-         retro_vkbd_transparent = !retro_vkbd_transparent;
-      }
-      else
-      if (vkflag[i] && (!(joypad_bits[0] & (1 << i)) &&
-                        !(joypad_bits[1] & (1 << i))))
-      {
-         vkflag[i] = 0;
-      }
-
-      /* Key press, RetroPad B joyports 1+2 / Keyboard Enter / Pointer */
-      i = RETRO_DEVICE_ID_JOYPAD_B;
-      if (!vkflag[i] && ((joypad_bits[0] & (1 << i)) ||
-                         (joypad_bits[1] & (1 << i)) ||
-                         input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_RETURN) ||
-                         input_state_cb(0, RETRO_DEVICE_POINTER, 0, RETRO_DEVICE_ID_POINTER_PRESSED)))
-      {
-         vkey_pressed = vkeys[(vkey_pos_y * VKBDX) + vkey_pos_x + ((retro_vkbd_page) ? VKBDX * VKBDY : 0)].value;
-
-         vkflag[i] = 1;
-
-         if (vkey_pressed == -1)
-            last_vkey_pressed = -1;
-         else if (vkey_pressed == -2)
-         {
-            last_vkey_pressed = -1;
-            retro_vkbd_page = !retro_vkbd_page;
-         }
-         else if (vkey_pressed < -10)
-         {
-            last_vkey_pressed = vkey_pressed;
-            last_vkey_pressed_time = now;
-
-            if (vkey_pressed == -11) /* Mouse up */
-            {
-               retro_mouse(0, 0, -3);
-               mflag[0][RETRO_DEVICE_ID_JOYPAD_UP] = 1;
-            }
-            else if (vkey_pressed == -12) /* Mouse down */
-            {
-               retro_mouse(0, 0, 3);
-               mflag[0][RETRO_DEVICE_ID_JOYPAD_DOWN] = 1;
-            }
-            else if (vkey_pressed == -13) /* Mouse left */
-            {
-               retro_mouse(0, -3, 0);
-               mflag[0][RETRO_DEVICE_ID_JOYPAD_LEFT] = 1;
-            }
-            else if (vkey_pressed == -14) /* Mouse right */
-            {
-               retro_mouse(0, 3, 0);
-               mflag[0][RETRO_DEVICE_ID_JOYPAD_RIGHT] = 1;
-            }
-            else if (vkey_pressed == -15) /* LMB */
-            {
-               retro_mouse_button(0, 0, 1);
-               mflag[0][RETRO_DEVICE_ID_JOYPAD_B] = 1;
-            }
-            else if (vkey_pressed == -16) /* RMB */
-            {
-               retro_mouse_button(0, 1, 1);
-               mflag[0][RETRO_DEVICE_ID_JOYPAD_A] = 1;
-            }
-            else if (vkey_pressed == -17) /* MMB */
-            {
-               retro_mouse_button(0, 2, 1);
-               mflag[0][RETRO_DEVICE_ID_JOYPAD_Y] = 1;
-            }
-            else if (vkey_pressed == -18) /* Toggle joystick/mouse */
-            {
-               emu_function(EMU_JOYMOUSE);
-            }
-            else if (vkey_pressed == -19) /* Toggle turbo fire */
-            {
-               emu_function(EMU_TURBO_FIRE);
-            }
-            else if (vkey_pressed == -20) /* Reset */
-            {
-               emu_function(EMU_RESET);
-            }
-            else if (vkey_pressed == -21) /* Toggle statusbar / save disk */
-            {
-               if (retro_capslock)
-                  emu_function(EMU_SAVE_DISK);
-               else
-                  emu_function(EMU_STATUSBAR);
-            }
-            else if (vkey_pressed == -22) /* Toggle aspect ratio / zoom mode */
-            {
-               if (retro_capslock)
-                  emu_function(EMU_ZOOM_MODE);
-               else
-                  emu_function(EMU_ASPECT_RATIO);
-            }
-         }
-         else
-         {
-            if (vkey_pressed == AK_CAPSLOCK)
-            {
-               retro_key_down(vkey_pressed);
-               retro_key_up(vkey_pressed);
-               retro_capslock = !retro_capslock;
-               last_vkey_pressed = -1;
-            }
-            else
-            {
-               last_vkey_pressed = vkey_pressed;
-               if (retro_capslock)
-                  retro_key_down(keyboard_translation[RETROK_LSHIFT]);
-
-               if (!vkey_sticky)
-               {
-                  if (vkey_pressed == vkey_sticky1)
-                     vkey_sticky1_release = 1;
-                  if (vkey_pressed == vkey_sticky2)
-                     vkey_sticky2_release = 1;
-               }
-               retro_key_down(vkey_pressed);
-            }
-         }
-      }
-      else
-      if (vkflag[i] && (!(joypad_bits[0] & (1 << i)) &&
-                        !(joypad_bits[1] & (1 << i)) &&
-                        !input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_RETURN) &&
-                        !input_state_cb(0, RETRO_DEVICE_POINTER, 0, RETRO_DEVICE_ID_POINTER_PRESSED)))
-      {
-         vkey_pressed = -1;
-         vkflag[i] = 0;
-
-         if (last_vkey_pressed > -10 && last_vkey_pressed < -15)
-         {
-            mspeed = mspeed_default;
-
-            if (vkey_pressed == -11) /* Mouse up */
-               mflag[0][RETRO_DEVICE_ID_JOYPAD_UP] = 0;
-            else if (vkey_pressed == -12) /* Mouse down */
-               mflag[0][RETRO_DEVICE_ID_JOYPAD_DOWN] = 0;
-            else if (vkey_pressed == -13) /* Mouse left */
-               mflag[0][RETRO_DEVICE_ID_JOYPAD_LEFT] = 0;
-            else if (vkey_pressed == -14) /* Mouse right */
-               mflag[0][RETRO_DEVICE_ID_JOYPAD_RIGHT] = 0;
-         }
-      }
-      else
-      if (vkflag[i] && ((joypad_bits[0] & (1 << i)) ||
-                        (joypad_bits[1] & (1 << i)) ||
-                        input_state_cb(0, RETRO_DEVICE_KEYBOARD, 0, RETROK_RETURN) ||
-                        input_state_cb(0, RETRO_DEVICE_POINTER, 0, RETRO_DEVICE_ID_POINTER_PRESSED)))
-      {
-         if (now - last_vkey_pressed_time > 100)
-         {
-            mspeed++;
-            last_vkey_pressed_time = now;
-         }
-
-         if (vkey_pressed == -11) /* Mouse up */
-            retro_mouse(0, 0, -mspeed);
-         else if (vkey_pressed == -12) /* Mouse down */
-            retro_mouse(0, 0, mspeed);
-         else if (vkey_pressed == -13) /* Mouse left */
-            retro_mouse(0, -mspeed, 0);
-         else if (vkey_pressed == -14) /* Mouse right */
-            retro_mouse(0, mspeed, 0);
-      }
-
-      if (vkflag[i] && vkey_pressed > 0)
-      {
-         if (let_go_of_button)
-            last_press_time_button = now;
-         if (now - last_press_time_button > VKBD_STICKY_HOLDING_TIME)
-            vkey_sticky = 1;
-         let_go_of_button = 0;
-      }
-      else
-      {
-         let_go_of_button = 1;
-         vkey_sticky = 0;
-      }
-   }
-#if 0
-   printf("vkey:%d sticky:%d sticky1:%d sticky2:%d, now:%d last:%d\n", vkey_pressed, vkey_sticky, vkey_sticky1, vkey_sticky2, now, last_press_time_button);
-#endif
+   input_vkbd();
 }
 
 int process_keyboard_pass_through()
@@ -2336,7 +1361,7 @@ void retro_poll_event()
    static int dpadmouse_speed[2] = {0};
    static int dpadmouse_pressed[2] = {0};
 #ifdef MOUSE_DPAD_ACCEL
-   static long now = 0;
+   long now = 0;
    now = retro_ticks() / 1000;
 #endif
